// src/lib/types.ts
export interface InspectionItem {
  id: string;
  category: string;
  name: string;
  status: string | undefined;
  remarks?: string;
}
// ...
export interface NetworkInspection {
  id: string;
  originalOfflineId?: string;
  createdAt: string;
  updatedAt: string;
  date?: string;
  time?: string;
  region: string;
  district: string;
  feederName: string;
  voltageLevel: string;
  referencePole: string;
  latitude: number;
  longitude: number;
  location?: string;
  status: "pending" | "in-progress" | "completed" | "rejected";
  inspector: {
    id: string;
    name: string;
    email: string;
    phone?: string;
  };
  // Pole Information
  poleId: string;
  poleHeight: "8m" | "9m" | "10m" | "11m" | "14m" | "others";
  poleType: "CP" | "WP" | "SP" | "ST";
  groundCondition: string;
  items: InspectionItem[];
  
  // Head Gears Information
  poleCondition: {
    tilted: boolean;
    rotten: boolean;
    burnt: boolean;
    substandard: boolean;
    conflictWithLV: boolean;
    notes: string;
  };
  
  stayCondition: {
    requiredButNotAvailable: boolean;
    cut: boolean;
    misaligned: boolean;
    defectiveStay: boolean;
    notes: string;
  };
  
  crossArmCondition: {
    misaligned: boolean;
    bend: boolean;
    corroded: boolean;
    substandard: boolean;
    others: boolean;
    notes: string;
  };
  
  insulatorCondition: {
    insulatorType?: string;
    brokenOrCracked: boolean;
    burntOrFlashOver: boolean;
    shattered: boolean;
    defectiveBinding: boolean;
    notes: string;
  };
  
  conductorCondition: {
    looseConnectors: boolean;
    weakJumpers: boolean;
    burntLugs: boolean;
    saggedLine: boolean;
    undersized: boolean;
    notes: string;
  };
  
  lightningArresterCondition: {
    brokenOrCracked: boolean;
    flashOver: boolean;
    noEarthing: boolean;
    bypassed: boolean;
    noArrester: boolean;
    notes: string;
  };
  
  dropOutFuseCondition: {
    brokenOrCracked: boolean;
    flashOver: boolean;
    insufficientClearance: boolean;
    looseOrNoEarthing: boolean;
    corroded: boolean;
    linkedHVFuses: boolean;
    others: boolean;
    notes: string;
  };
  
  transformerCondition: {
    leakingOil: boolean;
    lowOilLevel: boolean;
    missingEarthLeads: boolean;
    linkedHVFuses: boolean;
    rustedTank: boolean;
    crackedBushing: boolean;
    others: boolean;
    notes: string;
  };
  
  recloserCondition: {
    lowGasLevel: boolean;
    lowBatteryLevel: boolean;
    burntVoltageTransformers: boolean;
    protectionDisabled: boolean;
    bypassed: boolean;
    others: boolean;
    notes: string;
  };
  
  additionalNotes: string;
  images: string[];
  afterImages: string[];
}